# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentLabs REST API
  description: |
    Welcome to the AgentLabs REST API documentation. This API enables you to programmatically automate AI-powered voice communications at scale.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /agents:
    get:
      summary: List Agents
      description: |
        Retrieve all AI agents in your account. Each agent represents a configured voice assistant with its own personality, language, and behavior.
        
        <br />
        
        **Agent Status:**
        - `isActive: true` — Agent is ready for calls
        - `isActive: false` — Agent is disabled
        
        <br />
        
        **Required Scope:** `agents:read`
      operationId: listAgents
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of agents with basic information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/{id}:
    get:
      summary: Get Agent Details
      description: |
        Retrieve complete configuration for an agent including:
        - System prompt and first message
        - Voice and LLM settings
        - Transfer configuration
        - Language and personality settings
        
        <br />
        
        **Use Case:** Inspect agent settings before using in campaigns or calls.
        
        <br />
        
        **Required Scope:** `agents:read`
      operationId: getAgent
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Complete agent configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
...
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      allOf:
        - $ref: '#/components/schemas/Meta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                hasNext:
                  type: boolean
                hasPrev:
                  type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
        meta:
          $ref: '#/components/schemas/Meta'
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [flow, incoming]
        telephonyProvider:
          type: string
        language:
          type: string
        isActive:
          type: boolean
        transferEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentDetail:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            systemPrompt:
              type: string
            firstMessage:
              type: string
            llmModel:
              type: string
            temperature:
              type: number
            voiceId:
              type: string
            openaiVoice:
              type: string
            transferPhoneNumber:
              type: string
            maxDurationSeconds:
              type: integer
    AgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AgentDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AgentDetail'
        meta:
          $ref: '#/components/schemas/Meta'
...